refactor: api 호출 로직 분리
